Update to latest DR changes

+++ AgarPlugin/FoodItem.cs
@@ -4,7 +4,7 @@ namespace AgarPlugin
 {
     public class FoodItem : IEdible
     {
-        public uint ID { get; set; }
+        public ushort ID { get; set; }
         public float Radius => 0.1f;
         public float X { get; set; }
         public float Y { get; set; }
@@ -13,7 +13,7 @@ namespace AgarPlugin
         public byte ColorG { get; set; }
         public byte ColorB { get; set; }
 
-        public FoodItem(uint ID, float x, float y, byte colorR, byte colorG, byte colorB)
+        public FoodItem(ushort ID, float x, float y, byte colorR, byte colorG, byte colorB)
         {
             this.ID = ID;
             this.X = x;
